rdf class now uses the specified container to calculate distances and so obeys possible minimum image conventions

+++ classes/container.C
@@ -50,7 +50,7 @@ string cell::povray() {
   return o.str();
 }
 //----------- BOX --------------------------
-box::box(float sidelength) {
+box::box(double sidelength) {
   len = sidelength;
   len_half=len/2;
   len_inv=1/len;
